LPS-57869 renames files in social-group-statistics-web module
