+++ modules/apps/social/social-group-statistics-web/src/com/liferay/social/group/statistics/web/portlet/SocialGroupStatisticsPortlet.java
@@ -15,8 +15,8 @@
 package com.liferay.social.group.statistics.web.portlet;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.social.group.statistics.web.constants.GroupStatisticsPortletKeys;
-import com.liferay.social.group.statistics.web.upgrade.GroupStatisticsWebUpgrade;
+import com.liferay.social.group.statistics.web.constants.SocialGroupStatisticsPortletKeys;
+import com.liferay.social.group.statistics.web.upgrade.SocialGroupStatisticsWebUpgrade;
 
 import javax.portlet.Portlet;
 
@@ -40,18 +40,18 @@ import org.osgi.service.component.annotations.Reference;
 		"javax.portlet.expiration-cache=0",
 		"javax.portlet.init-param.template-path=/",
 		"javax.portlet.init-param.view-template=/view.jsp",
-		"javax.portlet.name=" + GroupStatisticsPortletKeys.GROUP_STATISTICS,
+		"javax.portlet.name=" + SocialGroupStatisticsPortletKeys.SOCIAL_GROUP_STATISTICS,
 		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=guest,power-user,user",
 		"javax.portlet.supports.mime-type=text/html"
 	},
 	service = Portlet.class
 )
-public class GroupStatisticsPortlet extends MVCPortlet {
+public class SocialGroupStatisticsPortlet extends MVCPortlet {
 
 	@Reference(unbind = "-")
-	protected void setGroupStatisticsWebUpgrade(
-		GroupStatisticsWebUpgrade groupStatisticsWebUpgrade) {
+	protected void setSocialGroupStatisticsWebUpgrade(
+		SocialGroupStatisticsWebUpgrade groupStatisticsWebUpgrade) {
 	}
 
 }
