+++ modules/apps/social/social-group-statistics-web/src/content/Language.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=column diagram
 group-statistics-chart-type-tag-cloud=tag cloud
 group-statistics-data-range-this-year=this year
 highest-activity-period=Highest Activity Period
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Group Statistics
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Group Statistics
 lowest-activity-period=Lowest Activity Period
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Please configure this portlet and select at least one activity counter.
 site-statistics=Site Statistics:
