+++ modules/apps/social/social-group-statistics-web/src/content/Language_ar.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=رسم العمود
 group-statistics-chart-type-tag-cloud=الدالات
 group-statistics-data-range-this-year=هذه السنة
 highest-activity-period=أكثر فترة نشاطا
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=إحصاء المجموعة
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=إحصاء المجموعة
 lowest-activity-period=أقل فترة نشاطا
 please-configure-this-portlet-and-select-at-least-one-activity-counter=يرجى تهيئة هذا المدخل واختيار عداد أنشطة واحد على الأقل.
 site-statistics=إحصائيات الموقع:
