+++ modules/apps/social/social-group-statistics-web/src/content/Language_ca.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=diagrama de columnes
 group-statistics-chart-type-tag-cloud=Núvol d'etiquetes
 group-statistics-data-range-this-year=Aquest any
 highest-activity-period=El període d'activitat més gran
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Estadístiques de grup
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Estadístiques de grup
 lowest-activity-period=Període d'activitat més baix
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Configureu aquest mòdul de portal i seleccioneu com a mínim un comptador d'activitats.
 site-statistics=Estadístiques del lloc:
