+++ modules/apps/social/social-group-statistics-web/src/content/Language_cs.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=sloupcový diagram
 group-statistics-chart-type-tag-cloud=mrak štítků
 group-statistics-data-range-this-year=tento rok
 highest-activity-period=Období nejvyšší aktivity
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Statistiky skupin
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Statistiky skupin
 lowest-activity-period=Období s nejnižší aktivitou
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Nastavte, prosím, tento portlet a vyberte alespoň jeden čítač aktivity.
 site-statistics=Statistiky webu:
