+++ modules/apps/social/social-group-statistics-web/src/content/Language_da.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=column diagram (Automatic Copy)
 group-statistics-chart-type-tag-cloud=Etiket-sky
 group-statistics-data-range-this-year=this year (Automatic Copy)
 highest-activity-period=Highest Activity Period (Automatic Copy)
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Gruppestatistik
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Gruppestatistik
 lowest-activity-period=Lowest Activity Period (Automatic Copy)
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Please configure this portlet and select at least one activity counter. (Automatic Copy)
 site-statistics=Site Statistics: (Automatic Copy)
