+++ modules/apps/social/social-group-statistics-web/src/content/Language_de.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=Spaltendiagramm
 group-statistics-chart-type-tag-cloud=Schlagwortwolke
 group-statistics-data-range-this-year=Dieses Jahr
 highest-activity-period=Zeitraum der größten Aktivität
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Gruppenstatistiken
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Gruppenstatistiken
 lowest-activity-period=Zeitraum der geringsten Aktivität
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Bitte konfigurieren Sie dieses Portlet und wählen Sie mindestens einen Aktivitätszähler aus.
 site-statistics=Statistiken der Site:
