+++ modules/apps/social/social-group-statistics-web/src/content/Language_el.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=διάγραμμα στηλών
 group-statistics-chart-type-tag-cloud=σύννεφο ετικεττών
 group-statistics-data-range-this-year=αυτό το έτος
 highest-activity-period=Περίοδος εντονότερης δραστηριότητας
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Στατιστικές ομάδων
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Στατιστικές ομάδων
 lowest-activity-period=Περίοδος χαμηλότερης δραστηριότητας
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Παρακαλώ ρυθμίστε αυτό το portlet και επιλέξτε τουλάχιστον έναν μετρητή δραστηριότητας
 site-statistics=Στατιστικά του site:
