+++ modules/apps/social/social-group-statistics-web/src/content/Language_es.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=Diagrama de columnas
 group-statistics-chart-type-tag-cloud=Nube de etiquetas
 group-statistics-data-range-this-year=este año
 highest-activity-period=El período más alto de actividad
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Estadísticas de Grupo
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Estadísticas de Grupo
 lowest-activity-period=El período más alto de actividad
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Por favor, configure este portlet y seleccione al menos un contador para las actividades.
 site-statistics=Estadísticas del sitio:
