+++ modules/apps/social/social-group-statistics-web/src/content/Language_eu.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=zutabe diagrama
 group-statistics-chart-type-tag-cloud=etiketen lainoa
 group-statistics-data-range-this-year=urte hau
 highest-activity-period=Jarduera Epe Handiena
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Talde Estatistikak
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Talde Estatistikak
 lowest-activity-period=Jarduera Epe Baxuena
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Mesedez, konfiguratu portlet hau eta aukeratu gitxienez jarduera zenbatzaile bat.
 site-statistics=Gunearen Estatistikak:
