+++ modules/apps/social/social-group-statistics-web/src/content/Language_fa.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=نمودار ستونی
 group-statistics-chart-type-tag-cloud=برچسب ابر
 group-statistics-data-range-this-year=امسال
 highest-activity-period=بیشترین مدت زمان فعالیت
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=آمار گروه
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=آمار گروه
 lowest-activity-period=حداقل مدت زمان فعالیت
 please-configure-this-portlet-and-select-at-least-one-activity-counter=لطفا این پرتلت را تنظیم کرده و حداقل یک شمارنده فعالیت انتخاب کنید.
 site-statistics=آمار سایت
