+++ modules/apps/social/social-group-statistics-web/src/content/Language_fi.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=sarakkeen kaaviokuva
 group-statistics-chart-type-tag-cloud=tagi pilvi
 group-statistics-data-range-this-year=tämä vuosi
 highest-activity-period=Korkein aktiviteetti ajanjakso
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Ryhmien tilastot
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Ryhmien tilastot
 lowest-activity-period=Hiljaisi aktiviteetti periodi
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Määritä asetukset tälle portletille ja valitse vähintää yksi aktiviteetti laskin.
 site-statistics=Sivuston statistiikka:
