+++ modules/apps/social/social-group-statistics-web/src/content/Language_fr.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=diagramme en colonnes
 group-statistics-chart-type-tag-cloud=Nuage d'étiquettes
 group-statistics-data-range-this-year=Cette année
 highest-activity-period=Période de plus grande activité
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Statistiques du groupe
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Statistiques du groupe
 lowest-activity-period=Période de plus basse activité
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Veuillez configurer cette portlet et sélectionner au moins un compteur d'activités.
 site-statistics=Statistiques du site :
