+++ modules/apps/social/social-group-statistics-web/src/content/Language_hi_IN.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=स्तंभ आरेख (Automat
 group-statistics-chart-type-tag-cloud=टॅग क्लाउड
 group-statistics-data-range-this-year=इस वर्ष (Automatic Translation)
 highest-activity-period=उच्चतम गतिविधि अवधि (Automatic Translation)
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=समूह के आँकड़े (Automatic Translation)
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=समूह के आँकड़े (Automatic Translation)
 lowest-activity-period=सबसे कम गतिविधि अवधि (Automatic Translation)
 please-configure-this-portlet-and-select-at-least-one-activity-counter=कृपया इस portlet को कॉन्फ़िगर करें और कम से कम एक गतिविधि काउंटर का चयन करें। (Automatic Translation)
 site-statistics=साइट आँकड़े: (Automatic Translation)
