+++ modules/apps/social/social-group-statistics-web/src/content/Language_hr.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=stupčasti dijagram
 group-statistics-chart-type-tag-cloud=Oblak tagova
 group-statistics-data-range-this-year=ova godina
 highest-activity-period=Period najveće aktivnosti
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Statistika grupa
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Statistika grupa
 lowest-activity-period=Period najmanje aktivnosti
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Molim konfigurirajte ovaj portlet i odaberite bar jedan brojač aktivnosti.
 site-statistics=Site Statistics: (Automatic Copy)
