+++ modules/apps/social/social-group-statistics-web/src/content/Language_hu.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=oszlopdiagram
 group-statistics-chart-type-tag-cloud=Címkefelhő
 group-statistics-data-range-this-year=ez az év
 highest-activity-period=Leginkább tevékeny időszak
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Csoportstatisztikák
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Csoportstatisztikák
 lowest-activity-period=Legalacsonyabb tevékenységi időszak
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Be kell állítani a portletet és kiválasztani legalább egy tevékenységszámlálót.
 site-statistics=Webhely statisztikák:
