+++ modules/apps/social/social-group-statistics-web/src/content/Language_in.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=kolom diagram
 group-statistics-chart-type-tag-cloud=lebel awan
 group-statistics-data-range-this-year=tahun ini
 highest-activity-period=Kegiatan Periode Tertinggi
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Kelompok Statistik
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Kelompok Statistik
 lowest-activity-period=Kegiatan Periode Terendah
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Silahkan mengkonfigurasi portlet ini dan pilih setidaknya satu counter aktivitas.
 site-statistics=Situs Statistik:
