+++ modules/apps/social/social-group-statistics-web/src/content/Language_it.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=grafico a colonne
 group-statistics-chart-type-tag-cloud=tag cloud
 group-statistics-data-range-this-year=quest'anno
 highest-activity-period=Periodo di Attività Maggiore
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Statistiche Gruppo
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Statistiche Gruppo
 lowest-activity-period=Periodo di Attività Minore
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Configura la portlet e seleziona almeno un contatore attività.
 site-statistics=Statistiche Sito:
