+++ modules/apps/social/social-group-statistics-web/src/content/Language_iw.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=תרשים עמודות
 group-statistics-chart-type-tag-cloud=ענן תגים
 group-statistics-data-range-this-year=השנה
 highest-activity-period=תקופת הפעילות הרבה ביותר
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=נתונים סטטיסטיים של קבוצה
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=נתונים סטטיסטיים של קבוצה
 lowest-activity-period=תקופת הפעילות הקצרה ביותר
 please-configure-this-portlet-and-select-at-least-one-activity-counter=אנא הגדר חלונית זו ובחר מונה פעילות אחד לפחות.
 site-statistics=נתונים סטטיסטיים של האתר:
