+++ modules/apps/social/social-group-statistics-web/src/content/Language_ja.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=コラムグラフ
 group-statistics-chart-type-tag-cloud=タグクラウド
 group-statistics-data-range-this-year=今年の情報を表示する
 highest-activity-period=最高活動期間
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=グループ統計
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=グループ統計
 lowest-activity-period=最低活動期間
 please-configure-this-portlet-and-select-at-least-one-activity-counter=このポートレットを設定し、一つ以上のアクティビティーカウンターを選択してください。
 site-statistics=サイト統計：
