+++ modules/apps/social/social-group-statistics-web/src/content/Language_ko.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=열 다이어그램 (Automatic Transl
 group-statistics-chart-type-tag-cloud=꼬리표 구름
 group-statistics-data-range-this-year=올해 (Automatic Translation)
 highest-activity-period=최고 활동 기간 (Automatic Translation)
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=그룹 통계 (Automatic Translation)
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=그룹 통계 (Automatic Translation)
 lowest-activity-period=가장 낮은 활동 기간 (Automatic Translation)
 please-configure-this-portlet-and-select-at-least-one-activity-counter=이 포틀릿 구성 하 고 하나 이상의 활동 카운터를 선택 하십시오. (Automatic Translation)
 site-statistics=사이트 통계: (Automatic Translation)
