+++ modules/apps/social/social-group-statistics-web/src/content/Language_lo.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=ຖັນແຜນວາດ
 group-statistics-chart-type-tag-cloud=ເທ໋ກຄູ
 group-statistics-data-range-this-year=ປີນີ້
 highest-activity-period=ໄລຍະເວລາກິດຈະກຳສູງສຸດ
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=ສະຖິຕິຂອງກຸ່ມ
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=ສະຖິຕິຂອງກຸ່ມ
 lowest-activity-period=ກຳນົດເວບລາກິດຈະກຳນ້ອຍສຸດ
 please-configure-this-portlet-and-select-at-least-one-activity-counter=ກະລຸນາຕິດຕັ້ງພ໋ອດເລ໋ດນີ້ ແລະເລືອກເປັນຢ່າງໜ້ອຍໜື່ງເຫດການການນັດ
 site-statistics=Site Statistics: (Automatic Copy)
