+++ modules/apps/social/social-group-statistics-web/src/content/Language_nb.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=kolonnediagram
 group-statistics-chart-type-tag-cloud=taggsky
 group-statistics-data-range-this-year=dette året
 highest-activity-period=Periode med høyest aktivitet
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Gruppestatistikk
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Gruppestatistikk
 lowest-activity-period=Perioden med lavest aktivitet
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Vennligst konfigurer dette portletet og velg mist en aktivitetsteller.
 site-statistics=Nettstedstatistikk:
