+++ modules/apps/social/social-group-statistics-web/src/content/Language_nl_BE.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=Kolommendiagram
 group-statistics-chart-type-tag-cloud=Labelwolk
 group-statistics-data-range-this-year=dit jaar
 highest-activity-period=Meest actieve periode
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Groepsstatistieken
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Groepsstatistieken
 lowest-activity-period=Periode met de minste activiteit
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Configureer deze portlet en selecteer minstens één activiteitensteller.
 site-statistics=Site statistieken: (Automatic Translation)
