+++ modules/apps/social/social-group-statistics-web/src/content/Language_pl.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=wykres słupkowy
 group-statistics-chart-type-tag-cloud=chmura tagów
 group-statistics-data-range-this-year=tym roku
 highest-activity-period=Najdłuższa aktywność
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Statystyki grupy
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Statystyki grupy
 lowest-activity-period=Okres najniższej aktywności
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Skonfiguruj ten portlet i wskaż co najmniej jeden licznik aktywności.
 site-statistics=Statystyki witryny:
