+++ modules/apps/social/social-group-statistics-web/src/content/Language_pt_BR.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=diagrama de coluna
 group-statistics-chart-type-tag-cloud=nuvem de tags
 group-statistics-data-range-this-year=este ano
 highest-activity-period=Período de maior atividade
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Estatísticas do Grupo
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Estatísticas do Grupo
 lowest-activity-period=Período de menor atividade
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Por favor configure este portlet e selecione pelo menos um contador de atividade.
 site-statistics=Estatísticas do site:
