+++ modules/apps/social/social-group-statistics-web/src/content/Language_pt_PT.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=diagrama de coluna (Automatic Transla
 group-statistics-chart-type-tag-cloud=Nuvem do Tag
 group-statistics-data-range-this-year=Este ano (Automatic Translation)
 highest-activity-period=Período de maior atividade (Automatic Translation)
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Estatísticas do Grupo
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Estatísticas do Grupo
 lowest-activity-period=Menor período de atividade (Automatic Translation)
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Por favor configure este portlet e selecione pelo menos uma contador de atividade. (Automatic Translation)
 site-statistics=Estatísticas do site: (Automatic Translation)
