+++ modules/apps/social/social-group-statistics-web/src/content/Language_ru.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=гистограмма
 group-statistics-chart-type-tag-cloud=Облако тегов
 group-statistics-data-range-this-year=в этом году
 highest-activity-period=Период наибольшей c активностью
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Статистика групп
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Статистика групп
 lowest-activity-period=Период наименьшей c активностью
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Пожалуйста, настройте этот портлет и выберите хотя бы один счетчик деятельности.
 site-statistics=Статистика сайта:
