+++ modules/apps/social/social-group-statistics-web/src/content/Language_sk.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=stĺpcový diagram
 group-statistics-chart-type-tag-cloud=oblak značiek
 group-statistics-data-range-this-year=tento rok
 highest-activity-period=Obdobie najvyššej aktivity
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Štatistiky skupiny
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Štatistiky skupiny
 lowest-activity-period=Obdobie s najnižšou aktivitou
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Prosím, nakonfigurujte tento portlet a vyberte aspoň jedno počítadlo aktivít.
 site-statistics=Štatistiky sídla:
