+++ modules/apps/social/social-group-statistics-web/src/content/Language_sr_RS.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=ступичасти дијаграм
 group-statistics-chart-type-tag-cloud=означен облак
 group-statistics-data-range-this-year=ова година
 highest-activity-period=Период Највише Активности
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Статистике Групе
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Статистике Групе
 lowest-activity-period=Период Најниже Активности
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Молимо вас да конфигуришете овај портлет и изаберете најмање једну бројачку активност.
 site-statistics=Site Statistics: (Automatic Copy)
