+++ modules/apps/social/social-group-statistics-web/src/content/Language_sr_RS_latin.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=stupičasti dijagram
 group-statistics-chart-type-tag-cloud=označen oblak
 group-statistics-data-range-this-year=ova godina
 highest-activity-period=Period Najviše Aktivnosti
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Statistike Grupe
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Statistike Grupe
 lowest-activity-period=Period Najniže Aktivnosti
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Molimo vas da konfigurišete ovaj portlet i izaberete najmanje jednu brojačku aktivnost.
 site-statistics=Site Statistics: (Automatic Copy)
