+++ modules/apps/social/social-group-statistics-web/src/content/Language_sv.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=stapeldiagram
 group-statistics-chart-type-tag-cloud=taggmoln
 group-statistics-data-range-this-year=detta året
 highest-activity-period=Perioden med högst aktivitet
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Gruppstatistik
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Gruppstatistik
 lowest-activity-period=Perioden med lägst aktivitet
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Please configure this portlet and select at least one activity counter. (Automatic Copy)
 site-statistics=Site Statistics: (Automatic Copy)
