+++ modules/apps/social/social-group-statistics-web/src/content/Language_vi.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=biểu đồ cột
 group-statistics-chart-type-tag-cloud=Đánh dấu (tag) đám mây
 group-statistics-data-range-this-year=năm nay
 highest-activity-period=Thời điểm hoạt động nhiều nhất
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=Thống kê nhóm người dùng
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=Thống kê nhóm người dùng
 lowest-activity-period=Thời kỳ thấp điểm
 please-configure-this-portlet-and-select-at-least-one-activity-counter=Vui lòng cấu hình ứng dụng (portlet) này và chọn ít nhất một bộ đếm đang hoạt động.
 site-statistics=Thống kê
