+++ modules/apps/social/social-group-statistics-web/src/content/Language_zh_CN.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=直方图
 group-statistics-chart-type-tag-cloud=标签云
 group-statistics-data-range-this-year=今年
 highest-activity-period=最高活跃期
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=群组统计
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=群组统计
 lowest-activity-period=最低的活跃期间
 please-configure-this-portlet-and-select-at-least-one-activity-counter=请配置此portlet并且选择至少一个活动计数器。
 site-statistics=站点统计：
