+++ modules/apps/social/social-group-statistics-web/src/content/Language_zh_TW.properties
@@ -13,7 +13,7 @@ group-statistics-chart-type-column-diagram=欄圖
 group-statistics-chart-type-tag-cloud=標籤雲
 group-statistics-data-range-this-year=今年
 highest-activity-period=最高的活動期間
-javax.portlet.title.com_liferay_social_group_statistics_web_portlet_GroupStatisticsPortlet=群組統計
+javax.portlet.title.com_liferay_social_group_statistics_web_portlet_SocialGroupStatisticsPortlet=群組統計
 lowest-activity-period=最低活動期間
 please-configure-this-portlet-and-select-at-least-one-activity-counter=請配置這Portlet並且至少選擇一個活動計數器。
 site-statistics=站台統計：
